requirements.txt
Django~=3.1
djangorestframework~=3.11.1
pandas~=1.1.1
django-cors-headers==3.5.0
yagmail==0.14.245
channels_redis==3.2.0